Sanitize data in list.js file

+++ server/controllers/list.js
@@ -1,3 +1,5 @@
+const sanitize = require('mongo-sanitize');
+
 const List = require('../models/list.model');
 const Item = require('../models/item.model');
 const {
@@ -37,7 +39,7 @@ const createList = (req, resp) => {
   });
 
   if (cohortId && !isListPrivate) {
-    Cohort.findOne({ _id: cohortId })
+    Cohort.findOne({ _id: sanitize(cohortId) })
       .then(cohort => {
         const { memberIds } = cohort;
 
@@ -81,23 +83,27 @@ const createList = (req, resp) => {
 };
 
 const deleteListById = (req, resp) => {
-  List.findOneAndDelete(
-    { _id: req.params.id, ownerIds: req.user._id },
-    (err, doc) => {
-      if (err) {
-        return resp.status(400).send({
-          message:
-            'An error occurred while deleting the list. Please try again.'
-        });
+  const {
+    user: { _id: userId },
+    params: { id: listId }
+  } = req;
+
+  List.findOneAndDelete({ _id: listId, ownerIds: userId })
+    .exec()
+    .then(doc => {
+      if (!doc) {
+        return resp.status(400).send({ message: 'List not found.' });
       }
 
-      doc
-        ? resp.status(200).send({
-            message: `List "${doc.name}" successfully deleted.`
-          })
-        : resp.status(404).send({ message: 'List not found.' });
-    }
-  );
+      return resp.status(200).send({
+        message: `List "${doc.name}" successfully deleted.`
+      });
+    })
+    .catch(() =>
+      resp.status(400).send({
+        message: 'An error occurred while deleting the list. Please try again.'
+      })
+    );
 };
 
 const getListsMetaData = (req, resp) => {
@@ -115,23 +121,23 @@ const getListsMetaData = (req, resp) => {
     query.cohortId = cohortId;
   }
 
-  List.find(
-    query,
-    '_id name description isPrivate items favIds cohortId',
-    { sort: { created_at: -1 } },
-    (err, docs) => {
-      if (err) {
-        return resp.status(400).send({
-          message:
-            'An error occurred while fetching the lists data. Please try again.'
-        });
+  List.find(query, '_id name description isPrivate items favIds cohortId', {
+    sort: { created_at: -1 }
+  })
+    .exec()
+    .then(docs => {
+      if (!docs) {
+        return resp.status(400).send({ message: 'No lists data found.' });
       }
 
-      docs
-        ? resp.status(200).json(responseWithLists(docs, userId))
-        : resp.status(404).send({ message: 'No lists data found.' });
-    }
-  );
+      return resp.status(200).json(responseWithLists(docs, userId));
+    })
+    .catch(() =>
+      resp.status(400).send({
+        message:
+          'An error occurred while fetching the lists data. Please try again.'
+      })
+    );
 };
 
 const getArchivedListsMetaData = (req, resp) => {
@@ -154,20 +160,24 @@ const getArchivedListsMetaData = (req, resp) => {
     `_id name description isPrivate items favIds isArchived ${
       cohortId ? 'cohortId' : ''
     }`,
-    { sort: { created_at: -1 } },
-    (err, docs) => {
-      if (err) {
-        return resp.status(400).send({
-          message:
-            'An error occurred while fetching the archived lists data. Please try again.'
-        });
+    { sort: { created_at: -1 } }
+  )
+    .exec()
+    .then(docs => {
+      if (!docs) {
+        return resp
+          .status(400)
+          .send({ message: 'No archived lists data found.' });
       }
 
-      docs
-        ? resp.status(200).json(responseWithLists(docs, userId))
-        : resp.status(404).send({ message: 'No archived lists data found.' });
-    }
-  );
+      return resp.status(200).json(responseWithLists(docs, userId));
+    })
+    .catch(() =>
+      resp.status(400).send({
+        message:
+          'An error occurred while fetching the archived lists data. Please try again.'
+      })
+    );
 };
 
 const addItemToList = (req, resp) => {
@@ -187,26 +197,27 @@ const addItemToList = (req, resp) => {
 
   List.findOneAndUpdate(
     {
-      _id: listId,
+      _id: sanitize(listId),
       $or: [{ ownerIds: userId }, { memberIds: userId }]
     },
     { $push: { items: item } },
-    { new: true },
-    (err, doc) => {
-      const newItem = doc.items.slice(-1)[0];
-
-      if (err) {
-        return resp.status(400).send({
-          message:
-            'An error occurred while adding a new item. Please try again.'
-        });
+    { new: true }
+  )
+    .exec()
+    .then(doc => {
+      if (!doc) {
+        return resp.status(400).send({ message: 'List  not found.' });
       }
 
-      doc
-        ? resp.status(200).send(responseWithItem(newItem, userId))
-        : resp.status(404).send({ message: 'List  not found.' });
-    }
-  );
+      const newItem = doc.items.slice(-1)[0];
+
+      return resp.status(200).send(responseWithItem(newItem, userId));
+    })
+    .catch(() =>
+      resp.status(400).send({
+        message: 'An error occurred while adding a new item. Please try again.'
+      })
+    );
 };
 
 const getListData = (req, resp) => {
@@ -308,29 +319,31 @@ const updateListItem = (req, resp) => {
   List.findOneAndUpdate(
     {
       _id: listId,
-      'items._id': itemId,
-      $or: [{ ownerIds: req.user._id }, { memberIds: req.user._id }]
+      'items._id': sanitize(itemId),
+      $or: [{ ownerIds: userId }, { memberIds: userId }]
     },
     {
       $set: dataToUpdate
     },
-    { new: true },
-    (err, doc) => {
-      if (err) {
-        return resp.status(400).send({
-          message:
-            'An error occurred while updating the list data. Please try again.'
-        });
+    { new: true }
+  )
+    .exec()
+    .then(doc => {
+      if (!doc) {
+        return resp.status(400).send({ message: 'List data not found.' });
       }
 
       const itemIndex = doc.items.findIndex(item => item._id.equals(itemId));
       const item = doc.items[itemIndex];
 
-      doc
-        ? resp.status(200).json(responseWithItem(item, userId))
-        : resp.status(404).send({ message: 'List data not found.' });
-    }
-  );
+      return resp.status(200).json(responseWithItem(item, userId));
+    })
+    .catch(() =>
+      resp.status(400).send({
+        message:
+          'An error occurred while updating the list data. Please try again.'
+      })
+    );
 };
 
 const voteForItem = (req, resp) => {
@@ -343,26 +356,28 @@ const voteForItem = (req, resp) => {
   List.findOneAndUpdate(
     {
       _id: listId,
-      'items._id': itemId,
-      $or: [{ ownerIds: req.user._id }, { memberIds: req.user._id }]
+      'items._id': sanitize(itemId),
+      $or: [{ ownerIds: userId }, { memberIds: userId }]
     },
-    { $push: { 'items.$.voterIds': req.user._id } },
-    { new: true },
-    (err, doc) => {
-      if (err) {
-        return resp.status(400).send({
-          message: 'An error occurred while voting. Please try again.'
-        });
+    { $push: { 'items.$.voterIds': userId } },
+    { new: true }
+  )
+    .exec()
+    .then(doc => {
+      if (!doc) {
+        return resp.status(400).send({ message: 'List data not found.' });
       }
 
       const itemIndex = doc.items.findIndex(item => item._id.equals(itemId));
       const item = doc.items[itemIndex];
 
-      doc
-        ? resp.status(200).json(responseWithItem(item, userId))
-        : resp.status(404).send({ message: 'List data not found.' });
-    }
-  );
+      return resp.status(200).json(responseWithItem(item, userId));
+    })
+    .catch(() =>
+      resp.status(400).send({
+        message: 'An error occurred while voting. Please try again.'
+      })
+    );
 };
 
 const clearVote = (req, resp) => {
@@ -375,30 +390,35 @@ const clearVote = (req, resp) => {
   List.findOneAndUpdate(
     {
       _id: listId,
-      'items._id': itemId,
-      $or: [{ ownerIds: req.user._id }, { memberIds: req.user._id }]
+      'items._id': sanitize(itemId),
+      $or: [{ ownerIds: userId }, { memberIds: userId }]
     },
-    { $pull: { 'items.$.voterIds': req.user._id } },
-    { new: true },
-    (err, doc) => {
-      if (err) {
-        return resp.status(400).send({
-          message: 'An error occurred while voting. Please try again.'
-        });
+    { $pull: { 'items.$.voterIds': userId } },
+    { new: true }
+  )
+    .exec()
+    .then(doc => {
+      if (!doc) {
+        return resp.status(400).send({ message: 'List data not found.' });
       }
 
       const itemIndex = doc.items.findIndex(item => item._id.equals(itemId));
       const item = doc.items[itemIndex];
 
-      doc
-        ? resp.status(200).json(responseWithItem(item, userId))
-        : resp.status(404).send({ message: 'List data not found.' });
-    }
-  );
+      return resp.status(200).json(responseWithItem(item, userId));
+    })
+    .catch(() =>
+      resp.status(400).send({
+        message: 'An error occurred while voting. Please try again.'
+      })
+    );
 };
 
 const updateListById = (req, resp) => {
   const { description, isArchived, name } = req.body;
+  const {
+    user: { _id: userId }
+  } = req;
   const { id: listId } = req.params;
   const dataToUpdate = filter(x => x !== undefined)({
     description,
@@ -409,145 +429,163 @@ const updateListById = (req, resp) => {
   List.findOneAndUpdate(
     {
       _id: listId,
-      $or: [{ ownerIds: req.user._id }]
+      $or: [{ ownerIds: userId }]
     },
-    dataToUpdate,
-    (err, doc) => {
-      if (err) {
-        return resp.status(400).send({
-          message:
-            'An error occurred while updating the list data. Please try again.'
-        });
+    dataToUpdate
+  )
+    .exec()
+    .then(doc => {
+      if (!doc) {
+        return resp.status(400).send({ message: 'List data not found.' });
       }
 
-      doc
-        ? resp
-            .status(200)
-            .send({ message: `List "${doc.name}" successfully updated.` })
-        : resp.status(404).send({ message: 'List data not found.' });
-    }
-  );
+      return resp
+        .status(200)
+        .send({ message: `List "${doc.name}" successfully updated.` });
+    })
+    .catch(() =>
+      resp.status(400).send({
+        message:
+          'An error occurred while updating the list data. Please try again.'
+      })
+    );
 };
 
 const addToFavourites = (req, resp) => {
   const { id: listId } = req.params;
+  const {
+    user: { _id: userId }
+  } = req;
 
   List.findOneAndUpdate(
     {
       _id: listId,
-      $or: [{ ownerIds: req.user._id }, { memberIds: req.user._id }]
+      $or: [{ ownerIds: userId }, { memberIds: userId }]
     },
     {
-      $push: { favIds: req.user._id }
-    },
-    (err, doc) => {
-      if (err) {
-        return resp.status(400).send({
-          message: "Can't mark list as favourite. Please try again."
-        });
+      $push: { favIds: userId }
+    }
+  )
+    .exec()
+    .then(doc => {
+      if (!doc) {
+        return resp.status(400).send({ message: 'List data not found.' });
       }
 
-      doc
-        ? resp.status(200).send({
-            message: `List "${doc.name}" successfully marked as favourite.`
-          })
-        : resp.status(404).send({ message: 'List data not found.' });
-    }
-  );
+      return resp.status(200).send({
+        message: `List "${doc.name}" successfully marked as favourite.`
+      });
+    })
+    .catch(() =>
+      resp.status(400).send({
+        message: "Can't mark list as favourite. Please try again."
+      })
+    );
 };
 
 const removeFromFavourites = (req, resp) => {
   const { id: listId } = req.params;
+  const {
+    user: { _id: userId }
+  } = req;
 
   List.findOneAndUpdate(
     {
       _id: listId,
-      $or: [{ ownerIds: req.user._id }, { memberIds: req.user._id }]
+      $or: [{ ownerIds: userId }, { memberIds: userId }]
     },
     {
-      $pull: { favIds: req.user._id }
-    },
-    (err, doc) => {
-      if (err) {
-        return resp.status(400).send({
-          message: "Can't remove list from favourites. Please try again."
-        });
+      $pull: { favIds: userId }
+    }
+  )
+    .exec()
+    .then(doc => {
+      if (!doc) {
+        return resp.status(400).send({ message: 'List data not found.' });
       }
 
-      doc
-        ? resp.status(200).send({
-            message: `List "${doc.name}" successfully removed from favourites.`
-          })
-        : resp.status(404).send({ message: 'List data not found.' });
-    }
-  );
+      return resp.status(200).send({
+        message: `List "${doc.name}" successfully removed from favourites.`
+      });
+    })
+    .catch(() =>
+      resp.status(400).send({
+        message: "Can't remove list from favourites. Please try again."
+      })
+    );
 };
 
 const removeOwner = (req, resp) => {
   const { id: listId } = req.params;
   const { userId } = req.body;
   const {
-    user: { _id: ownerId }
+    user: { _id: currentUserId }
   } = req;
 
   List.findOneAndUpdate(
-    { _id: listId, ownerIds: { $all: [ownerId, userId] } },
-    { $pull: { ownerIds: userId, memberIds: userId, viewersIds: userId } },
-    (err, doc) => {
-      if (err) {
-        return resp.status(400).send({
-          message: "Can't remove owner from list."
-        });
+    { _id: listId, ownerIds: { $all: [currentUserId, sanitize(userId)] } },
+    { $pull: { ownerIds: userId, memberIds: userId, viewersIds: userId } }
+  )
+    .exec()
+    .then(doc => {
+      if (!doc) {
+        return resp.status(400).send({ message: 'List data not found.' });
       }
 
-      if (doc) {
-        return resp.status(200).send({
-          message: 'Owner successfully removed from list.'
-        });
-      }
-
-      resp.status(404).send({ message: 'List data not found.' });
-    }
-  );
+      return resp.status(200).send({
+        message: 'Owner successfully removed from list.'
+      });
+    })
+    .catch(() =>
+      resp.status(400).send({
+        message: "Can't remove owner from list."
+      })
+    );
 };
 
 const removeMember = (req, resp) => {
   const { id: listId } = req.params;
   const { userId } = req.body;
   const {
-    user: { _id: ownerId }
+    user: { _id: currentUserId }
   } = req;
+  const sanitizedUserId = sanitize(userId);
 
   List.findOneAndUpdate(
-    { _id: listId, ownerIds: ownerId, viewersIds: userId },
-    { $pull: { viewersIds: userId, memberIds: userId, ownerIds: userId } },
-    (err, doc) => {
-      if (err) {
-        return resp.status(400).send({
-          message: "Can't remove member from list."
-        });
+    { _id: listId, ownerIds: currentUserId, viewersIds: sanitizedUserId },
+    {
+      $pull: {
+        viewersIds: sanitizedUserId,
+        memberIds: sanitizedUserId,
+        ownerIds: sanitizedUserId
       }
-
-      if (doc) {
-        return resp.status(200).send({
-          message: 'Member successfully removed from list.'
-        });
-      }
-
-      resp.status(404).send({ message: 'List data not found.' });
     }
-  );
+  )
+    .exec()
+    .then(doc => {
+      if (!doc) {
+        return resp.status(400).send({ message: 'List data not found.' });
+      }
+
+      return resp.status(200).send({
+        message: 'Member successfully removed from list.'
+      });
+    })
+    .catch(() =>
+      resp.status(400).send({
+        message: "Can't remove member from list."
+      })
+    );
 };
 
 const addOwnerRole = (req, resp) => {
   const { id: listId } = req.params;
   const { userId } = req.body;
-
   const {
-    user: { _id: ownerId }
+    user: { _id: currentUserId }
   } = req;
 
-  List.findOne({ _id: listId, ownerIds: ownerId })
+  List.findOne({ _id: listId, ownerIds: currentUserId })
     .populate('cohortId', 'memberIds ownerIds')
     .exec()
     .then(doc => {
@@ -627,13 +665,10 @@ const addMemberRole = (req, resp) => {
   const { id: listId } = req.params;
   const { userId } = req.body;
   const {
-    user: { _id: ownerId }
+    user: { _id: currentUserId }
   } = req;
 
-  List.findOne({
-    _id: listId,
-    ownerIds: { $in: [ownerId] }
-  })
+  List.findOne({ _id: listId, ownerIds: { $in: [currentUserId] } })
     .exec()
     .then(doc => {
       if (!doc) {
@@ -674,13 +709,10 @@ const removeMemberRole = (req, resp) => {
   const { id: listId } = req.params;
   const { userId } = req.body;
   const {
-    user: { _id: ownerId }
+    user: { _id: currentUserId }
   } = req;
 
-  List.findOne({
-    _id: listId,
-    ownerIds: { $in: [ownerId] }
-  })
+  List.findOne({ _id: listId, ownerIds: { $in: [currentUserId] } })
     .populate('cohortId', 'memberIds ownerIds viewersIds')
     .exec()
     .then(doc => {
@@ -703,11 +735,7 @@ const removeMemberRole = (req, resp) => {
 
       return doc.save();
     })
-    .then(() =>
-      resp.status(200).send({
-        message: 'User has no member role'
-      })
-    )
+    .then(() => resp.status(200).send({ message: 'User has no member role' }))
     .catch(err => {
       if (err instanceof BadRequestException) {
         const { status, message } = err;
@@ -737,7 +765,7 @@ const addViewer = (req, resp) => {
     .then(doc => {
       list = doc;
 
-      return User.findOne({ email }).exec();
+      return User.findOne({ email: sanitize(email) }).exec();
     })
     .then(userData => {
       if (!userData) {
@@ -786,12 +814,15 @@ const updateItemDetails = (req, resp) => {
   const { description, link, itemId } = req.body;
   const { id: listId } = req.params;
   const dataToUpdate = updateSubdocumentFields('items', { description, link });
+  const {
+    user: { _id: userId }
+  } = req;
 
   List.findOneAndUpdate(
     {
       _id: listId,
-      'items._id': itemId,
-      $or: [{ ownerIds: req.user._id }, { memberIds: req.user._id }]
+      'items._id': sanitize(itemId),
+      $or: [{ ownerIds: userId }, { memberIds: userId }]
     },
     { $set: dataToUpdate },
     (err, doc) => {
@@ -820,7 +851,7 @@ const cloneItem = (req, resp) => {
 
   List.findOne({
     _id: listId,
-    'items._id': itemId,
+    'items._id': sanitize(itemId),
     $or: [{ ownerIds: userId }, { memberIds: userId }]
   })
     .exec()
